requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openpyxl>=3.1.0
python-docx>=1.1.0
